Changed cost in items table to extend to 3 decimal points. Aircraft spruce has their cost go out to 3 decimal points for item cost. When multiple of same item is ordered it matters. Single item is rounded.

+++ php/activation.php
@@ -26,7 +26,7 @@ function purchase_records_activation(){
 		build_category_id int(10) unsigned NOT NULL,
 		istool bit(1) DEFAULT b'0',
 		item varchar(64) NOT NULL,
-		cost double(10,2) NOT NULL DEFAULT 0.00,
+		cost double(10,3) NOT NULL DEFAULT 0.00,
 		quantity int(10) unsigned NOT NULL DEFAULT 1,
 		weblink varchar(128) DEFAULT '',
 		PRIMARY KEY  (purchase_id),
